Rename script to download_cache_process
